requirements.txt
python-telegram-bot==20.7
openai==1.12.0
fastapi==0.109.2
uvicorn==0.27.1
gunicorn==21.2.0
python-dotenv==1.0.1
httpx==0.26.0
pydantic==2.6.1
typing-extensions==4.9.0 